+++ src/main/java/ch/heigvd/bdr/controllers/ResourceControllerInterface.java
@@ -6,13 +6,13 @@ import java.sql.SQLException;
 import io.javalin.http.Context;
 
 public interface ResourceControllerInterface {
-  void all(Context ctx) throws ClassNotFoundException, IOException, SQLException;
+    void all(Context ctx) throws ClassNotFoundException, IOException, SQLException;
 
-  void show(Context ctx) throws ClassNotFoundException, IOException, SQLException;
+    void show(Context ctx) throws ClassNotFoundException, IOException, SQLException;
 
-  void create(Context ctx) throws ClassNotFoundException, IOException, SQLException;
+    void create(Context ctx) throws ClassNotFoundException, IOException, SQLException;
 
-  void update(Context ctx) throws ClassNotFoundException, IOException, SQLException;
+    void update(Context ctx) throws ClassNotFoundException, IOException, SQLException;
 
-  void delete(Context ctx) throws ClassNotFoundException, IOException, SQLException;
+    void delete(Context ctx) throws ClassNotFoundException, IOException, SQLException;
 }
