+++ src/main/java/ch/heigvd/bdr/misc/StringHelper.java
@@ -3,6 +3,7 @@ package ch.heigvd.bdr.misc;
 public class StringHelper {
     /**
      * Check if a string contains an integer
+     *
      * @param str
      * @return
      */
