+++ src/main/java/ch/heigvd/bdr/models/CollaboratorNeed.java
@@ -4,25 +4,25 @@ package ch.heigvd.bdr.models;
  * Stores all the data related to the collaborator needs of a task
  */
 public class CollaboratorNeed {
-  private UserRole type;
-  private int quantity;
+    private UserRole type;
+    private int quantity;
 
     public CollaboratorNeed() {
     }
 
-  public CollaboratorNeed(UserRole type, int quantity) {
-    this.type = type;
-    this.quantity = quantity;
-  }
+    public CollaboratorNeed(UserRole type, int quantity) {
+        this.type = type;
+        this.quantity = quantity;
+    }
 
-  // Getters and setters
-  public UserRole getType() {
-    return type;
-  }
+    // Getters and setters
+    public UserRole getType() {
+        return type;
+    }
 
-  public void setType(UserRole type) {
-    this.type = type;
-  }
+    public void setType(UserRole type) {
+        this.type = type;
+    }
 
     public int getQuantity() {
         return quantity;
@@ -32,12 +32,12 @@ public class CollaboratorNeed {
         this.quantity = quantity;
     }
 
-  public static UserRole fromInt(int i) {
-    for (UserRole collaborator : UserRole.values()) {
-      if (collaborator.getValue() == i) {
-        return collaborator;
-      }
+    public static UserRole fromInt(int i) {
+        for (UserRole collaborator : UserRole.values()) {
+            if (collaborator.getValue() == i) {
+                return collaborator;
+            }
+        }
+        throw new IllegalArgumentException("Unexpected value: " + i);
     }
-    throw new IllegalArgumentException("Unexpected value: " + i);
-  }
 }
