+++ src/main/java/ch/heigvd/bdr/models/Material.java
@@ -4,16 +4,16 @@ package ch.heigvd.bdr.models;
  * Used to define available material
  */
 public enum Material {
-  LICENSE(0),
-  SERVER(1),
-  DATABASE(2);
-  private final int value;
+    LICENSE(0),
+    SERVER(1),
+    DATABASE(2);
+    private final int value;
 
-  Material(int value) {
-    this.value = value;
-  }
+    Material(int value) {
+        this.value = value;
+    }
 
-  public int getValue() {
-    return value;
-  }
+    public int getValue() {
+        return value;
+    }
 }
