+++ src/main/java/ch/heigvd/bdr/models/Project.java
@@ -4,40 +4,41 @@ package ch.heigvd.bdr.models;
  * Stores all the data related a project
  */
 public class Project {
-  private int id;
-  private String name;
-  private String description;
+    private int id;
+    private String name;
+    private String description;
 
-  public Project(){ }
+    public Project() {
+    }
 
-  public Project(int id, String name, String description) {
-    this.id = id;
-    this.name = name;
-    this.description = description;
-  }
+    public Project(int id, String name, String description) {
+        this.id = id;
+        this.name = name;
+        this.description = description;
+    }
 
-  // Getters and setters
-  public int getId() {
-    return id;
-  }
+    // Getters and setters
+    public int getId() {
+        return id;
+    }
 
-  public void setId(int id) {
-    this.id = id;
-  }
+    public void setId(int id) {
+        this.id = id;
+    }
 
-  public String getName() {
-    return name;
-  }
+    public String getName() {
+        return name;
+    }
 
-  public void setName(String name) {
-    this.name = name;
-  }
+    public void setName(String name) {
+        this.name = name;
+    }
 
-  public String getDescription() {
-    return description;
-  }
+    public String getDescription() {
+        return description;
+    }
 
-  public void setDescription(String description) {
-    this.description = description;
-  }
+    public void setDescription(String description) {
+        this.description = description;
+    }
 }
