+++ src/main/java/ch/heigvd/bdr/models/TaskDeadline.java
@@ -4,5 +4,5 @@ package ch.heigvd.bdr.models;
  * Used to define the tasks duration
  */
 public enum TaskDeadline {
-  THREE_MONTHS, ONE_YEAR, THREE_YEARS
+    THREE_MONTHS, ONE_YEAR, THREE_YEARS
 }
