+++ src/main/java/ch/heigvd/bdr/models/TaskPriority.java
@@ -4,5 +4,5 @@ package ch.heigvd.bdr.models;
  * Used to define a priority to a task
  */
 public enum TaskPriority {
-  LOW, MEDIUM, HIGH
+    LOW, MEDIUM, HIGH
 }
