+++ src/main/java/ch/heigvd/bdr/models/User.java
@@ -6,62 +6,63 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
  * Stores all the data related to a user
  */
 public class User {
-  private int id;
-  private String firstname;
-  private String lastname;
-  private String email;
-  private UserRole role;
-  @JsonIgnore
-  private String password;
+    private int id;
+    private String firstname;
+    private String lastname;
+    private String email;
+    private UserRole role;
+    @JsonIgnore
+    private String password;
 
-  public User(){}
+    public User() {
+    }
 
-  public User(int id, String firstname, String lastname, String email, UserRole role) {
-    this.id = id;
-    this.firstname = firstname;
-    this.lastname = lastname;
-    this.email = email;
-    this.role = role;
-  }
+    public User(int id, String firstname, String lastname, String email, UserRole role) {
+        this.id = id;
+        this.firstname = firstname;
+        this.lastname = lastname;
+        this.email = email;
+        this.role = role;
+    }
 
-  // Getters and setters
-  public int getId() {
-    return id;
-  }
+    // Getters and setters
+    public int getId() {
+        return id;
+    }
 
-  public void setId(int id) {
-    this.id = id;
-  }
+    public void setId(int id) {
+        this.id = id;
+    }
 
-  public String getFirstname() {
-    return firstname;
-  }
+    public String getFirstname() {
+        return firstname;
+    }
 
-  public void setFirstname(String firstname) {
-    this.firstname = firstname;
-  }
+    public void setFirstname(String firstname) {
+        this.firstname = firstname;
+    }
 
-  public String getLastname() {
-    return lastname;
-  }
+    public String getLastname() {
+        return lastname;
+    }
 
-  public void setLastname(String lastname) {
-    this.lastname = lastname;
-  }
+    public void setLastname(String lastname) {
+        this.lastname = lastname;
+    }
 
-  public String getEmail() {
-    return email;
-  }
+    public String getEmail() {
+        return email;
+    }
 
-  public void setEmail(String email) {
-    this.email = email;
-  }
+    public void setEmail(String email) {
+        this.email = email;
+    }
 
-  public UserRole getRole() {
-    return role;
-  }
+    public UserRole getRole() {
+        return role;
+    }
 
-  public void setRole(UserRole role) {
-    this.role = role;
-  }
+    public void setRole(UserRole role) {
+        this.role = role;
+    }
 }
