+++ src/main/java/ch/heigvd/bdr/models/UserRole.java
@@ -4,20 +4,20 @@ package ch.heigvd.bdr.models;
  * Used to manage the user's roles on the platform
  */
 public enum UserRole {
-  ADMIN(0),
-  MANAGER(1),
-  CONTRIBUTOR(2),
-  DEVELOPER(3),
-  SCRUM_MASTER(4),
-  DATA_SPECIALIST(5);
+    ADMIN(0),
+    MANAGER(1),
+    CONTRIBUTOR(2),
+    DEVELOPER(3),
+    SCRUM_MASTER(4),
+    DATA_SPECIALIST(5);
 
-  private final int value;
+    private final int value;
 
-  UserRole(int value) {
-    this.value = value;
-  }
+    UserRole(int value) {
+        this.value = value;
+    }
 
-  public int getValue() {
-    return value;
-  }
+    public int getValue() {
+        return value;
+    }
 }
